fix: check git and cd cmd when init clonning project

+++ src/util.ts
@@ -97,6 +97,15 @@ export async function copyRecursive(source: string, destination: string, exclude
   }
 }
 
+export function isGitInstalled(): boolean {
+  try {
+    execSync('git --version');
+    return true;
+  } catch (error) {
+    return false;
+  }
+}
+
 export function gitCloneAndDownloadFolderSync(
   repoUrl: string,
   branch: string,
@@ -106,6 +115,11 @@ export function gitCloneAndDownloadFolderSync(
   console.log('start cloning the dapp template..');
   const tempFolder = path.resolve(dappTemplatePath, 'temp-clone-folder');
 
+  if (!isGitInstalled()) {
+    console.log('Git is not installed, please check https://git-scm.com/');
+    return process.exit(1);
+  }
+
   // Empty the temp folder if it exists
   if (fs.existsSync(tempFolder)) {
     fs.rmSync(tempFolder, { recursive: true });
@@ -114,14 +128,22 @@ export function gitCloneAndDownloadFolderSync(
   // Create the temp folder
   fs.mkdirSync(tempFolder, { recursive: true });
 
-  const cloneCommand = `
-git clone -n --depth=1 --single-branch --branch ${branch} --filter=tree:0 \
-${repoUrl} ${tempFolder}
-cd ${tempFolder}
-git sparse-checkout set ${subFolderName}
-git checkout
-`;
-  execSync(cloneCommand);
+  // Clone the repository
+  try {
+    const cloneCommand = `git clone -n --depth=1 --single-branch --branch ${branch} --filter=tree:0 ${repoUrl} ${tempFolder}`;
+    execSync(cloneCommand);
+  } catch (error) {
+    console.error('Error:', error);
+    process.exit(1);
+  }
+  // checkout the examples sub folder
+  try {
+    execSync(`git sparse-checkout set ${subFolderName}`, { cwd: tempFolder });
+    execSync(`git checkout`, { cwd: tempFolder });
+  } catch (error) {
+    console.error('Error:', error);
+    process.exit(1);
+  }
 
   // Ensure targetPath exists and is a directory
   if (!fs.existsSync(targetPath) || !fs.statSync(targetPath).isDirectory()) {
